피보나치2

+++ 백준/3053/3053.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main(){
-	float n;
+	double n;
 	cin >> n;
 	
 	printf("%.6f\n",n*n*M_PI);
